misc code cleanup

+++ matsim/src/test/java/org/matsim/facilities/ActivityFacilitiesImplTest.java
@@ -33,7 +33,7 @@ public class ActivityFacilitiesImplTest {
 	void testAddActivityFacility() {
 		ActivityFacilities facilities = FacilitiesUtils.createActivityFacilities();
 		ActivityFacilitiesFactory factory = facilities.getFactory();
-		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord((double) 200, (double) 5000));
+		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord(200, 5000));
 
 		Assertions.assertEquals(0, facilities.getFacilities().size());
 
@@ -41,7 +41,7 @@ public class ActivityFacilitiesImplTest {
 
 		Assertions.assertEquals(1, facilities.getFacilities().size());
 
-		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord((double) 300, (double) 4000));
+		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord(300, 4000));
 		facilities.addActivityFacility(facility2);
 
 		Assertions.assertEquals(2, facilities.getFacilities().size());
@@ -51,8 +51,8 @@ public class ActivityFacilitiesImplTest {
 	void testAddActivityFacility_addingTwice() {
 		ActivityFacilities facilities = FacilitiesUtils.createActivityFacilities();
 		ActivityFacilitiesFactory factory = facilities.getFactory();
-		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord((double) 200, (double) 5000));
-		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord((double) 300, (double) 4000));
+		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord(200, 5000));
+		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord(300, 4000));
 
 		Assertions.assertEquals(0, facilities.getFacilities().size());
 
@@ -72,8 +72,8 @@ public class ActivityFacilitiesImplTest {
 	void testAddActivityFacility_sameId() {
 		ActivityFacilities facilities = FacilitiesUtils.createActivityFacilities();
 		ActivityFacilitiesFactory factory = facilities.getFactory();
-		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord((double) 200, (double) 5000));
-		ActivityFacility facility2 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord((double) 300, (double) 4000));
+		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord(200, 5000));
+		ActivityFacility facility2 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord(300, 4000));
 
 		Assertions.assertEquals(0, facilities.getFacilities().size());
 
@@ -95,8 +95,8 @@ public class ActivityFacilitiesImplTest {
 	void testRemove() {
 		ActivityFacilities facilities = FacilitiesUtils.createActivityFacilities();
 		ActivityFacilitiesFactory factory = facilities.getFactory();
-		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord((double) 200, (double) 5000));
-		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord((double) 300, (double) 4000));
+		ActivityFacility facility1 = factory.createActivityFacility(Id.create(1, ActivityFacility.class), new Coord(200, 5000));
+		ActivityFacility facility2 = factory.createActivityFacility(Id.create(2, ActivityFacility.class), new Coord(300, 4000));
 		facilities.addActivityFacility(facility1);
 		facilities.addActivityFacility(facility2);
 		Assertions.assertEquals(2, facilities.getFacilities().size());
